+++ bankfeeds/src/sdk/models/operations/deletebankfeedcredentials.ts
@@ -6,21 +6,15 @@ import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import * as shared from "../shared";
 import { AxiosResponse } from "axios";
 
-export class ProxyRequest extends SpeakeasyBase {
+export class DeleteBankFeedCredentialsRequest extends SpeakeasyBase {
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
-
-    /**
-     * The endpoint and query parameters made to the underlying integration.
-     */
-    @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=endpoint" })
-    endpoint: string;
 }
 
-export class ProxyResponse extends SpeakeasyBase {
+export class DeleteBankFeedCredentialsResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -29,12 +23,6 @@ export class ProxyResponse extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
-
-    /**
-     * Success
-     */
-    @SpeakeasyMetadata()
-    proxyResponse?: Record<string, any>;
 
     @SpeakeasyMetadata()
     statusCode: number;