+++ bankfeeds/RELEASES.md
@@ -657,3 +657,13 @@ Based on:
 - [typescript v0.35.0] bankfeeds
 ### Releases
 - [NPM v0.35.0] https://www.npmjs.com/package/@codat/bank-feeds/v/0.35.0 - bankfeeds
+
+## 2023-08-31 00:04:28
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.75.0 (2.89.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.36.0] bankfeeds
+### Releases
+- [NPM v0.36.0] https://www.npmjs.com/package/@codat/bank-feeds/v/0.36.0 - bankfeeds